backend/requirements.txt
fastapi
uvicorn
firebase-admin
python-dotenv
openai
pydantic
requests
safetensors
numpy

fastapi
uvicorn
firebase-admin
requests
python-dotenv


fastapi
uvicorn
firebase-admin
requests
python-dotenv
